Race condition in resume_foreground(DispatcherQueue)
It's possible for the lambda to run before control
returns to the caller of TryEnqueue. This led to async_suspend
accessing m_queued after it has already been destroyed
by the lambda.

We make the lambda responsible for setting m_queued prior to
resuming the coroutine.  The lambda can infer that the task
was queued by the fact that the lambda is running at all!

If the enqueue fails, then m_queued remains at its initial
value of false.