Race condition in resume_foreground(DispatcherQueue)
It's possible for the lambda to run before control
returns to the caller of TryEnqueue. This led to async_suspend
accessing m_queued after it has already been destroyed.

We need to prevent the lambda from invoking the handler until
async_suspend has stored the result of TryEnqueue into m_queued.
We do this by making the lambda wait until async_suspend is ready
before invoking the handler. The slim_event must live in the
awaiter (rather than as a local in await_suspend) because the
lambda may run after await_suspend has returned.

I chose a slim_event rather than a full kernel event because
allocating a kernel event for every co_await seemed too expensive,
especially because the wait is almost never going to block.

This commit introduces winrt::slim_event which is an event that
starts out unsignaled but can be signaled. Once signaled, it
cannot be reset. It is the size of a bool. For simplicity, only
infinite waits are permitted. (Otherwise, I would have to recalculate
the timeout when a spurious wake occurs.)