Do not pass nullptr arrays to the ABI
For array parameters, ensure that nullptr is not passed
as the array base, even for empty arrays. If the array
is empty, the array base is not used, but RPC still requires
it to be a non-null pointer.

We use a properly aligned non-null pointer as our fake
array base, in case somebody cares about alignment.

Note that this fix does not address [out] parameters.
The workaround for [out] parameters is more expensive
(need to CoTaskMemAlloc(0)) and it is necessary only
for out-of-process calls, so I don't want to incur the
overhead for everybody. People returning empty com_arrays
will have to remain vigilant.

Also fix BO in test Class when asked to fill an empty array.