Improve performance of make_factory with static_lifetime
Previous code constructed a new object each time,
even though only the first call will actually need
the constructed object. The second and subsequent
calls just throw the object away and return the
existing object.

Construct the object only if we have a reasonable chance
of actually using it. This is a big win for static-lifetime
objects that are expensive to construct.