Detect RPC server crash when co_await'ing
During a co_await of an IAsyncAction or IAsyncOperation,
if the completion delegate is released without ever being
invoked, then assume that the remote server crashed.
Resume the coroutine, and the RPC error will be raised
when await_resume calls GetResults().

Note that this behavior applies only to co_await.
If you manually attach a Completed handler, then
it is your responsibility to detect server death.

The disconnect_aware_handler is basically the same
as the lambda we had been using, just with the additional
behavior of resuming the coroutine if it is destructed
without ever having been invoked. We null out the coroutine
handle on invoke. That lets us detect the error case
in our destructor.