Add Sdk.props/targets to NuGet package
This is a crude, but effective, workaround for the lack of
PackageReference support in vcxproj files. With this change
applied, users can reference the package using an <Sdk />
element in the project (or an imported props/targets file).
This does not require packages.config, and the package will
be restored into the global NuGet cache.