Improve async cancellation interop
Other projections report cancellation via the Status property
and Completed delegate, rather than by GetResults. Switch from
GetResults to Status and Completed to detect cancellation,
to improve interop.

To avoid unnecessary fetches of the Status property
(which are virtual calls), we cache the status passed
to the Completed delegate. This applies to `co_await`,
`get()`, and `when_any()`. `when_all` is built out of
`co_await`, so it gets cancellation interop for free.

Factored cancellation checking into helper function
`check_status_canceled`. And factored "create a delegate but give me
access to the functor that the delegate is wrapping" into helper
function `make_delegate_with_shared_state`. This saves an allocation
in `when_any`. Would be nice if this functionality were exposed
in the public API surface, since it's a handy pattern.

`wait_for_completed` now returns the status, so the caller can avoid a
virtual call to the Status property. There is a tiny data race in
`wait_for_completed` if the operation completes immediately after our
wait timed out. In that case, we may return `Started` even though the
operation has completed, but this is an inherent race condition
(TOCTTOU), so it's unavoidable. Unfortunately, I couldn't figure out
an easy way to get rid of the language-lawyer data race.
Changing `status` to a `std::atomic` doesn't work because `std::atomic`
is not movable. And `_InterlockedExchange_nf` is supported only on ARM64.

Having `wait_for_completed` return the status allows `wait_for`
to avoid a virtual call to the Status property.

Fixes #641