Don't destroy the IAsyncInfo from inside Completed handler
If the Completed handler runs synchronously, that would
cause us to destroy the operation while a method on
the operation is still running. That's not very nice.

This stems from the fact that synchronous completion
causes the coroutine to resume before put_Completed
returns, and resuming the coroutine causes it to get
the result and then destroy the async operation.
To avoid this problem, we extend the lifetime of the
async operation inside await_suspend, so that the
synchronous completion doesn't destruct the async operation
until after put_Completed has returned.

If the operation completes asynchronously, then the
destruction happens at coroutine resumption, which is fine.
The put_Completed method returned a long time ago.

We cannot "steal" the async operation into await_suspend,
because await_resume needs the operation to get the results.
We have to make a copy. This costs us an AddRef/Release pair,
but avoiding the AddRef/Release will cost us a few atomic
operations (and memory barriers), so it's basically a wash.