Improve error message if you co_await a non-awaitable
Previous error message was

```
error C2672: 'get_awaiter': no matching overloaded function found
```

New error message is

```
error C2338: Not an awaitable type
```

If you `co_await` something that has both a member `operator co_await`
as well as a free `operator co_await`, the error message used to be

```
error C2668: 'winrt::impl::get_awaiter': ambiguous call to overloaded function
```

and it is now

```
error C2338: Ambiguous operator co_await (as both member and free function)
```

The C++ standard has rules for resolving the ambiguity, but they are
difficult to express in template metaprogramming, so we don't try.
Hasn't been a problem yet. Added some ambiguous awaiters to the unit
tests to verify that we resolve to the correct one.

Remove vestigial support for free awaiters, which is left over from
an earlier iteration of the coroutine TS.

(There is an interesting asymmetry in awaiter_finder: We always search
for `await_ready` on an lvalue, whereas the `operator co_await` uses the
original reference class of the object being awaited. This is a
consequence of the language specification, which converts the awaiter to
an lvalue *after* applying `operator co_await` but *before* calling
`await_ready`.)