Make test reports deterministic in the face of nondeterminism
The concurrency tests for multi-threaded collections
run for a nondeterministic number of iterations. This means
that each time you run them, the numbers in the message

```
All tests passed (314159 assertions in 2718 test cases)
```

are different. This does not contribute to confidence that the
tests are working.

There is also a problem with the REQUIRE macro: It does not support
concurrent usage from multiple threads. If two threads perform a
REQUIRE at the same time, the assertion count is unreliable.

Fix both problems by not using REQUIRE (1) in nondeterministic ways
or (2) from multiple threads simultaneously. Instead, we use
`check_bool`, which will still throw an exception and fail the test,
but does not run afoul of REQUIRE's limitations.