Status() doesn't need a lock
Reading the status of an IAsyncXxx does not require
a lock because the operation is already inherently
racy. Change the `m_status` to a `std::atomic` so we
can access it from multiple threads without a lock.
Everybody accesses it as `std::memory_order_relaxed`
because they either are already protected by a lock
(so memory ordering is already protected) or they are
the `Status()` method which can read a value even
though the lock is taken by another thread that could
be changing the status.

Reading the value while another thread is changing it
is okay, as long as we get either the entire old value
or the entire new value, which is what `std::memory_order_relaxed`
gives us. If we read a stale value, that's okay, because
multithreading means that we could have executed a few
cycles sooner, in which case the stale value was current
at the time.

If anybody makes a decision based on the status, they
will have to return to the IAsyncXxx to get the Result
or the ErrorCode, and those will take the lock and
ensure consistency of the other protected members.

This change takes a lock out of the `await_transform`
which would otherwise be entered (and exited)
at every `co_await` operation. The code sequence

```
    lea      rcx, this->m_mutex
    call     AcquireSRWLockExclusive
    mov      rdi, this->m_status
    lea      rcx, this->m_mutex
    call     ReleaseSRWLockExclusive
    cmp      rdi, 2
    jz       throw_cancelled
```

becomes

```
    cmp      this->m_status, 2
    jz       throw_cancelled
```

which avoids two calls and frees up a nonvolatile register.

Relaxed order access to atomics is implemented as a simple load
or store on all supported platforms, so the change from a simple
variable to a `std::atomic` has no code generation effect, but it
is required to keep the compiler happy.