Use instrumented vectors/maps to force race conditions more efficiently
Instead of running 100,000 iterations hoping to hit a race condition,
use an instrumented vector/map that lets us inject delays into specific
operations, and which detects concurrency violations. This allows us
to force specific race conditions rather than waiting for them to happen.

This reduces the running time of the multi* tests from 80 seconds to
1.5 seconds, as measured on a Surface Pro 5 (dual Core-i5 7300U).

You can flip the #if at the top of multi_threaded_vector.cpp and
multi_threaded_map.cpp to force the test to use a single-threaded
collection and observe that the race conditions are hit.