get_error_token() allows a coroutine to simulate noexcept
There is until now no way to make a coroutine noexcept,
which makes "fail fast on exception" harder to debug.
The closest you can get is

```cpp
IAsyncAction DoSomethingAsync() try
{
    TryIt();
}
catch (...)
{
    winrt::terminate(); // fail fast on all unhandled exceptions
}
```

however this does not generate a good crash dump stack because
the `catch (...)` handles the exception, so it unwinds. All
you know is that *somebody* in your coroutine raised an exception.
The stowed exception gives you a stack trace, but you really
want to be able to inspect local variables, but those have
already unwound and are no longer available.

We introduce the "error token" which lets you configure the
coroutine's error behavior.

```cpp
IAsyncAction DoSomethingAsync()
{
    auto error = co_await get_error_token();
    error.set_noexcept(); // unhandled exceptions are now fatal

    TryIt(); // if this throws an exception, we will fail fast
}
```

This time, an exception from `TryIt()` will fail fast with
the exception-throwing frame still on the stack and therefore
captured in the dump for post-mortem analysis.

The `set_noexcept()` method takes a boolean parameter (default true)
that specifies whether the coroutine is logically noexcept.
It returns the previous value, in case you want to restore it later.