Multithreading tests: Add dummy definition to keep compiler happy
This type is instantiated by the deleted constructor whose original
purpose is to allow an implicit conversion from non-const iterator to
const iterator. When applied to the const iterator itself, it
enables an implicit conversion from garbage to const iterator.
The conversion is deleted, but the compiler still instantiates it,
so we define the type to keep everybody happy.