Make return address extraction work on more compilers
At a minimum, avoid build breaks on non-MSVC compilers.

Turns out that gcc, clang, icc, djgpp, ellcc, and zapcc
all define the symbol `__GNUC__` and support the gcc return address
extraction extensions. So we need only test two symbols:
`_MSC_VER` for MSVC, and `__GNUC__` for everyone else.

(clang and icc implement the gcc extensions for compatibility.
djgpp is built on top of gcc, so it inherits the gcc extensions.
ellcc and zapcc are built on top of clang, so they inherit the
gcc extensions via clang.)