Always bounce through the threadpool when switching from STA
Previously, we changed the way we switch from STA to NA by
bouncing through the threadpool so that we don't run the NA
on an STA thread.

We extend this behavior to switches from STA to STA, to avoid
RO_E_BLOCKED_CROSS_ASTA_CALL exceptions.

Interestingly, we add this support by deleting code. We would
normally add tests for STA and MAINSTA to the NA test, but since
MTA was ruled out by the previous test, and NA+STA+MAINSTA completely
cover all the non-MTA possibilities, there's no need to perform any
test at all. The fact that we got here means that the destination
is NA/STA/MAINSTA.

Note also that the first test in this block checks if we are resuming
in the same apartment that we are currently in. That test is
important, because it avoids us bouncing through the threadpool
just to return to our own thread, which would otherwise burn a threadpool
thread.

Added a new unit test to verify that STA-to-STA transfers release
the source STA. The test failed before the change, and passes after.

Note however that this introduces a possibly undesirable behavior change:
Suppose STA thread "A" hangs and STA thread "B" has many tasks, some of
which want to resume on STA thread "A".

Under the old behavior, STA thread "B" would run the first task,
and then that task will try to switch to STA thread "A" and hang.
STA thread "B" would remain hung and be unavailable to process tasks.
When STA thread "A" finally wakes up, STA thread "B" also wakes up
and resumes doing work. The number of hung threads is capped at the
number of STA threads.

Under the new behavior, STA thread "B" would run the first task,
and then ask a threadpool to resume on STA thread "A". The threadpool
thread hangs instead of STA thread "B". STA thread "B" is now available
to do other work (yay), but at a cost of a hung threadpool thread.
One of the things STA thread "B" might do is create more work items
that want to switch to STA thread "A", so we slowly (or maybe quickly)
exhaust the threadpool.

Mitigation for this worst-case scenario is that we sort of had that
problem anyway: If any task on a threadpool thread wants to switch to
STA thread "A", it will hang with or without these changes. Since
it is common to pass through a threadpool thread at some point,
you are likely to exhaust the threadpool even under the old algorithm.