Allow empty apartment_contexts
Constructing an apartment_context with nullptr creates an
empty apartment context that cannot be used for anything,
but you can later assign another apartment context to it.

This is handy if you don't want to capture the apartment
context right now, but you have to declare the variable
right now. For example, member variables are constructed
when the containing class is constructed, even though
you may not want to capture the apartment context until
later.

Prior to this fix, you had to use std::optional to defer
the construction of the apartment_context.

It's unfortunate that adding std::move support is so wordy.

We take advantage of this in the unit test itself: The
code to generate an apartment_context from a IContextCallback,
for example, needs to predeclare an empty apartment_context
so that it can be filled in from inside the IContextCallback
callback function.

An empty apartment_context is represented by an apartment type
of -1, which is not a legal value: It corresponds to
APTTYPE_CURRENT, which is never returned by CoGetApartmentType.