Factories no longer default to no_weak_ref
Originally, factories defaulted to no_weak_ref because there wasn't
any use case for weak references to factories, and the thought was
that the weak reference support would be a performance penalty.

In practice, the weak reference support does not have a significant
performance impact, so it's okay to leave it on even for factories.

Allowing weak references to factories is important because factories
implement the static members of a runtime class, and those members
may require taking weak references:

* Event handlers need weak references so that the event handler
  doesn't create a circular reference back to the factory.
* Queueing work to a background thread takes a weak reference
  so that the work can be abandoned if the factory has been destroyed.
  (We don't want to extend the lifetime of COM objects beyond the
  lifetime of COM itself.)

Furthermore, the ability to take a weak reference to a factory
is required in order to use event_revoker for static events,
since event_revoker requires that the event source support weak
references.