Improve diagnostics when implements<> is used incorectly
A common error is forgetting to provide the name of the derived
class as the first `implements<>` template parameter. This results
in a strange `implements_type is not defined` error. Detect this
case with a `static_assert` with a more help error message.

Note that we have to put the `static_assert` inside the implementation
of `GetRuntimeClassName` rather than putting it in the class definition
of `implements`. That's because `implements` is sometimes used with
an incomplete type, and we cannot use `is_base_of` with incomplete types.
Wait until `GetRuntimeClassName` is generated, at which point the class
is complete.