Properly report failure to switch to target apartment
If we were unable to switch to a target apartment
(either by an explicit `co_await apartment_context`
or implicitly at the completion of ` co_await IAsyncInfo`),
we now report the exception in the `co_await`'ing coroutine,
so it can be caught and handled.

Note that when this error occurs, we are stuck in limbo.
We left the original apartment for the MTA, and now we can't
get from the MTA to the target apartment. We have no choice but
to raise the exception on the MTA, even though the coroutine
may not have expected to be running there.

```cpp
IAsyncAction Trapped()
{
    Object o; // some object with a destructor
    co_await destroyed_ui_context; // throws an exception
}
```

The destructor for `Object` will run on the MTA, even though
the `Trapped` coroutine never runs on the MTA under normal
conditions.

This is making the best of a bad situation. We need to raise
the exception into the coroutine so that the coroutine can
enter a failure state and allow the coroutine chain to proceed.

Since the object `o` already had to be prepared for destruction
on the original apartment or the switched-to apartment (in case
an exception occurs in either apartment), we know that the
object cannot have apartment affinity anyway. So we're probably
okay to destruct it on the MTA.

The old code raised the exception on whatever thread happened
to notice that something bad happened. Sometimes that's okay,
for example if you do a `co_await destroyed_ui_context` from
an MTA thread, in which case the exception is thrown into the
coroutine and can be caught, or left uncaught and send the
coroutine into an error state.

However, all of the other cases weren't quite so happy.
To avoid deadlocks, the apartment switch is routed through
the threadpool, and that means that the failure is raised
on a threadpool thread. There's nobody around to catch that
exception, so the process fails fast.

We solve the problem by having an atomic hresult in the
awaiter, which is initialized to 0 (`S_OK`) and transitions
to error when we are about to resume the coroutine on the wrong
apartment, so that the error can be propagated during
`await_resume()`, which happens in the context of the awaiting
coroutine and therefore can be caught by the coroutine
(or sent to `unhandled_exception`).

The hresult in which we record the failure needs to be atomic
to avoid data races in case multiple threads detect failure
simultaneously. However, the access to it can be
relaxed, because it is set and consumed by the same thread.

The current implementation pulls a sneaky trick: To avoid
having to create an `operator co_await` for the `apartment_context`
object, we continue to let it be its own awaiter. This means
that if the apartment switch fails, we record the failure in
the `apartment_context` itself. We are making the assumption
that once an apartment context goes bad, it is unrecoverably
broken.